requirements.txt
tensorflow==2.11.0
gradio
numpy==1.23.5
Pillow
